feat: bootstrap CLI
This commit splits existing cmd into `cli` and `controlplane` to let
users communicate with controlplane (API) by CLI.
