+++ internal/cmd/cli/root.go
@@ -1,4 +1,4 @@
-package cmd
+package cli
 
 import (
 	"context"
@@ -19,9 +19,5 @@ var (
 )
 
 func Execute(ctx context.Context) error {
-	rootCmd.AddCommand(newAPICommand())
-	rootCmd.AddCommand(newAPIProxyCommand())
-	rootCmd.AddCommand(newControllerCommand())
-	rootCmd.AddCommand(newVersionCommand())
 	return rootCmd.ExecuteContext(ctx)
 }
