+++ internal/cmd/controlplane/api_proxy.go
@@ -1,26 +1,26 @@
-package cmd
+package controlplane
 
 import (
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
-	"github.com/akuity/kargo/internal/api/server"
+	"github.com/akuity/kargo/internal/api/proxy"
 	"github.com/akuity/kargo/internal/config"
 	"github.com/akuity/kargo/internal/logging"
 )
 
-func newAPICommand() *cobra.Command {
+func newAPIProxyCommand() *cobra.Command {
 	return &cobra.Command{
-		Use:               "api",
+		Use:               "api-proxy",
 		DisableAutoGenTag: true,
 		SilenceErrors:     true,
 		SilenceUsage:      true,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			cfg := config.NewAPIConfig()
+			cfg := config.NewAPIProxyConfig()
 			logger := log.New()
 			logger.SetLevel(cfg.LogLevel)
 			ctx := logging.ContextWithLogger(cmd.Context(), logger.WithFields(nil))
-			srv := server.NewServer(cfg)
+			srv := proxy.NewServer(cfg)
 			return srv.Serve(ctx)
 		},
 	}
