+++ internal/cmd/controlplane/controller.go
@@ -1,4 +1,4 @@
-package cmd
+package controlplane
 
 import (
 	argocd "github.com/argoproj/argo-cd/v2/pkg/apis/application/v1alpha1"
