+++ internal/cmd/controlplane/root.go
@@ -1,0 +1,27 @@
+package controlplane
+
+import (
+	"context"
+
+	"github.com/spf13/cobra"
+)
+
+var (
+	rootCmd = &cobra.Command{
+		Use:               "kargo",
+		DisableAutoGenTag: true,
+		SilenceErrors:     true,
+		SilenceUsage:      true,
+		Run: func(cmd *cobra.Command, args []string) {
+			cmd.HelpFunc()(cmd, args)
+		},
+	}
+)
+
+func Execute(ctx context.Context) error {
+	rootCmd.AddCommand(newAPICommand())
+	rootCmd.AddCommand(newAPIProxyCommand())
+	rootCmd.AddCommand(newControllerCommand())
+	rootCmd.AddCommand(newVersionCommand())
+	return rootCmd.ExecuteContext(ctx)
+}
